Updated tests for CalculationServicesTest and CriticalDamageBonusCalculator.

+++ src/main/java/Tekiz/_DPSCalculator/_DPSCalculator/model/enums/modifiers/ModifierTypes.java
@@ -14,7 +14,7 @@ public enum ModifierTypes
 	RANGE(Double.class),
 	SNEAK_DAMAGE(Double.class),
 	CRITICAL(Double.class),
-	CRITICAL_CONSUMPTION(Double.class),
+	CRITICAL_CONSUMPTION(Integer.class),
 	HEALTH(Double.class),
 	SPECIAL_STRENGTH(Integer.class),
 	SPECIAL_PERCEPTION(Integer.class),
